Updated Code to reflect Refactoring
