+++ demo/demo.py
@@ -5,12 +5,12 @@ from logsmith import log
 
 from providers.authLib.authLib import AUTHLIB
 from providers.authLib.standards.return_codes import RETURN_CODES as AUTHLIB_RC
-from providers.loginProvider.login import LOGIN
-from providers.loginProvider.standards.return_codes import RETURN_CODES as LRC
+from providers.login.login import LOGIN
+from providers.login.standards.return_codes import RETURN_CODES as LRC
 from providers.mailer.content import CONTENT_FACTORY
 from providers.mailer.mailer import MAILER
-from providers.registrationProvider.registration import REGISTER
-from providers.registrationProvider.standards.return_codes import RETURN_CODES as RRC
+from providers.registration.registration import REGISTER
+from providers.registration.standards.return_codes import RETURN_CODES as RRC
 
 # Setup Logsmith
 log = log()
